:recycle: (search-service) delegate aggregation tasks to specific services inheriting from generic search.service.ts
